refactor: 型付け専用スキーマをアプリ中核の実 DDL へ昇格させ本番の実型に合わせる
sqlc の型付け専用だった queries/schema.sql を infra/database/schema/core.sql へ移し、
そのまま起動時に流せる DDL にする（CREATE TABLE / INDEX を IF NOT EXISTS に、
採番列を bigserial に、これまで AutoMigrate が作っていた索引を明記）。
ナレッジ基盤の 2 ファイルと同じく、宣言（sqlc）と実体（DDL）を 1 ファイルに集約する。

あわせて、本番の information_schema と突き合わせて食い違っていた型を実型に合わせた。

- roles.id / users.role_id: smallint → integer（本番・AutoMigrate とも integer）
- company_applications の 4 列: text → varchar（200 / 120 / 255 / 16）
- users.role_id に DEFAULT 3（ローリングデプロイ中の旧コード向けの安全弁）

roles.id の型変更で sqlc の Go 型が int16 → int32 になるため、呼び出し側も追随させる。

+++ backend/internal/adapter/persistence/sqlcgen/company_stats.sql.go
@@ -32,7 +32,7 @@ type CountMembersByCompanyRow struct {
 // trainee 判定は正規化後の正である role_id で行う（パラメータで渡す）。
 // WHERE で company_id IS NOT NULL に絞っているので COALESCE は NULL を返さない
 // （型を非 NULL の bigint に確定させ、詰め替えを綺麗にするための cast）。
-func (q *Queries) CountMembersByCompany(ctx context.Context, traineeRoleID int16) ([]CountMembersByCompanyRow, error) {
+func (q *Queries) CountMembersByCompany(ctx context.Context, traineeRoleID int32) ([]CountMembersByCompanyRow, error) {
 	rows, err := q.db.QueryContext(ctx, countMembersByCompany, traineeRoleID)
 	if err != nil {
 		return nil, err
